refactor(test): split testCaseReader tests into focused files
Split test/util/testCaseReader.test.ts (1227 lines, 57 tests) into 7 focused test files:
- standaloneTestsFile.test.ts (20 tests) - Tests for readStandaloneTestsFile
- singleTestCase.test.ts (6 tests) - Tests for readTest
- multipleTestCases.test.ts (23 tests) - Tests for readTests
- csvRowConversion.test.ts (1 test) - Tests for testCaseFromCsvRow
- varsFileReader.test.ts (2 tests) - Tests for readTestFiles
- globPatternLoader.test.ts (1 test) - Tests for loadTestsFromGlob
- csvJsonFieldParsing.test.ts (4 tests) - Tests for CSV parsing with JSON fields

Changes:
- Created test/util/testCaseReader/ subdirectory for better organization
- Each file tests a single function/concern
- Improved test names to be action-oriented without 'should'
- Added proper setup/teardown with beforeEach/afterEach
- Fixed mock isolation issues for randomized test execution
- Cleaned up unnecessary mocks from each file
- Fixed JavaScript module tests by mocking importModule instead of virtual mocks
- Added README.md documenting the test organization
- All 56 tests now pass (previously 3 were skipped)

BREAKING CHANGE: None - this is a test-only refactor
